Changed Version to decimal.

+++ Commands/Information/CmdLastCmd.cs
@@ -30,7 +30,7 @@ namespace CommandDll
         public string Name { get { return "Last Command"; } }
         public CommandTypes Type { get { return CommandTypes.information; } }
         public string Author { get { return "Arrem"; } }
-        public int Version { get { return 1; } }
+        public decimal Version { get { return 1.00m; } }
         public string CUD { get { return ""; } }
         public byte Permission { get { return 80; } }
